% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripted-utilities.R
\name{check_dependencies}
\alias{check_dependencies}
\title{Load a check file and check dependencies.}
\usage{
check_dependencies(job, file, logger)
}
\arguments{
\item{file}{path to check file}

\item{logger}{logger for output}
}
\value{
list with hash for each dependency
}
\description{
Load a check file and check dependencies.
}
