% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripted-utilities.R
\name{check_file_path}
\alias{check_file_path}
\title{Create a .check file name for the script}
\usage{
check_file_path(target_dir, stub)
}
\arguments{
\item{target_dir}{directory to look in.}
}
\value{
single-string character vector of the file name.
}
\description{
This is not a fail-safe procedure.
}
