% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripted-utilities.R
\name{create_check_file}
\alias{create_check_file}
\title{Create a .check file for the script}
\usage{
create_check_file(target_dir, job, logger)
}
\arguments{
\item{target_dir}{directory to look in}
}
\value{
check file path
}
\description{
This is a .yaml file that records file content checksums
and their modification times, etc... to verify whether a file
is available and has been updated.
}
