% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruction-utilities.R
\name{create_target}
\alias{create_target}
\title{Create target dir, etc...}
\usage{
create_target(job, target, logger)
}
\arguments{
\item{target}{directory where output will go.}

\item{logger}{how to log}
}
\description{
Create target dir, etc...
}
