% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruction-utilities.R
\name{load_script}
\alias{load_script}
\title{Load a script}
\usage{
load_script(script, debug, logger)
}
\arguments{
\item{script}{path to script}

\item{debug}{whether to echo during source}
}
\description{
Load a script
}
