% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\docType{class}
\name{output-class}
\alias{output-class}
\alias{store}
\title{A wrapper class so that objects returned can save themselves.}
\description{
A wrapper class so that objects returned can save themselves.
}
\section{Fields}{

\describe{
\item{\code{format}}{format to save in}

\item{\code{data}}{data to save access function}

\item{\code{.data}}{actual data to save}

\item{\code{.logger}}{function to log with..}
}}


