% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-rdump.R
\name{stan_rdump}
\alias{stan_rdump}
\title{Straight from stan-dev/rstan}
\usage{
stan_rdump(list, file = "", append = FALSE, envir = parent.frame(),
  width = options("width")$width, quiet = FALSE)
}
\description{
link: https://github.com/stan-dev/rstan/blob/9949c1982ce47911b454484a526716a5cb9f9a32/rstan/rstan/R/misc.R
}
